The avatars show with different size #140 *We started off by introducing a new CSS class called '.height-remover' with 'height: revert-layer' as its main property. However, we hit a wall - 'revert-layer' is supported only by Firefox, which made it unsuitable for our cross-browser needs.
We then turned to JavaScript, thinking we could dynamically adjust the height.
This solution did introduce some lag and the issue persisted in edit mode, with the 'style' attribute disappearing.

Finally, on our third attempt, we made changes to the Velocity macro and added a new style named '.fixedSize-SIZE'.
This solution worked effectively in both view and edit modes without causing any lag.